+++ src/main/java/de/maxhenkel/audioplayer/api/events/GetDistanceEvent.java
@@ -9,6 +9,8 @@ public interface GetDistanceEvent {
 
     float getDefaultDistance();
 
+    float getItemDistance();
+
     void setDistance(float distance);
 
     float getDistance();